# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/me/password": {
      "patch": {
        "operationId": "UsersController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Current password incorrect",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change current user password",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/orders/{orderNumber}/items/{itemId}/review": {
      "post": {
        "operationId": "ReviewsController_createReview",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Item already reviewed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Review a purchased order item (order must be completed)",
        "tags": [
          "Reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "BW Café Ordering API",
    "description": "Single-restaurant delivery/pickup ordering API — guest and registered checkout, admin order management.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 409
          },
          "message": {
            "type": "string",
            "example": "Conflict"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/auth/register/request-otp"
          }
        },
        "required": [
          "code",
          "message",
          "timestamp",
          "path"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "data": {
            "type": "object",
            "nullable": true,
            "additionalProperties": false,
            "example": null
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data",
          "error"
        ]
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Ahmed Mohamed"
          },
          "phone": {
            "type": "string",
            "example": "+22200000000"
          },
          "role": {
            "type": "string",
            "example": "USER",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "status": {
            "type": "string",
            "example": "ACTIVE",
            "enum": [
              "PENDING_VERIFICATION",
              "ACTIVE",
              "BLOCKED"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "role",
          "status"
        ]
      },
      "UserProfileDataDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "UserProfileResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/UserProfileDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 150,
            "description": "Display name (2–150 chars)",
            "example": "Ahmed Ali"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "minLength": 8,
            "maxLength": 64,
            "description": "Current password",
            "example": "OldSecret123"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "maxLength": 64,
            "pattern": "/^(?=.*[A-Za-z])(?=.*\\d).+$/",
            "description": "New password — must contain letters and numbers (8–64 chars)",
            "example": "NewSecret456"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
      "SuccessDataDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        },
        "required": [
          "message"
        ]
      },
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/SuccessDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5,
            "example": 5
          },
          "comment": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000,
            "example": "Great coffee!"
          }
        },
        "required": [
          "rating"
        ]
      },
      "ReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "productId": {
            "type": "number",
            "example": 3
          },
          "rating": {
            "type": "number",
            "example": 5,
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string",
            "nullable": true,
            "example": "Great coffee!"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-07-13T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "productId",
          "rating",
          "createdAt"
        ]
      },
      "ReviewDataDto": {
        "type": "object",
        "properties": {
          "review": {
            "$ref": "#/components/schemas/ReviewDto"
          }
        },
        "required": [
          "review"
        ]
      },
      "ReviewResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ReviewDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
    }
  }
}